requirements.txt
wget
pandas
pytz
bs4
google-api-python-client
google-auth-httplib2
google-auth-oauthlib